backend/requirements.txt
huggingface-hub
strands-agents
strands-agents-tools
requests>=2.32.0
urllib3>=2.0,<3
chardet>=3.0.2,<6.0.0
crawl4ai
ddgs
fastapi
transformers
diffusers
accelerate
# starlette_context # for windows or linux
